pyproject.toml
[project]
name = "proyecto-alex"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Alcázar",email = "aalcazarprogramacion@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<5.3)",
    "pillow (>=12.2.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
